+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.0] - (27.10.2022)
+## Changes
+- added Power Motion Sensor for use with automations
+- added Volume Motion Sensor for use with automations (every volume change report motion)
+- added Mute Motion Sensor for use with automations
+- config.schema updated
+- other small fixes
+
 ## [3.13.32] - (10.09.2022)
 ## Changes
 - bump dependencies